Lots of changes going on right now
I had to add the JSon API JSR 353. You might want to remove it in
environments that provide it (like most app servers)

Some package reorganizations

The registration process is a stateful one (we need to check if the
response matches the request). I still want to keep the server
stateless. So I had to externalize the process state
(RegistrationState). And I use the challenge parameter as an identifier
for this object. Google seems to use a separate session id. But that
seems just to be yet another randomly chosen parameter, like the
challenge. I might be wrong. But the specs are not very clear about the
communication between the client browser and the server. But you can
still bind that state object to the user's session and there you have
your session parameter. 

One more example why the communication between the browser and the
server really need clarification:

The "FIDO U2F JavaScript API" document states that the RegisterResponse
has the following attributes:
* version
* registrationData
* clientData

Firefox just sends registrationData and clientData. Chrome sends all
three + challenge + keys (I hope it's RegisteredKeys as in the sign
request but there is no trace of it in any doc). I don't want to do the
JSON parsing/binding myself. After all, that's what these JAXB derivates
like resteasy and jackson are for. 

I'm still wondering how this appID/orgin thing is supposed to work.
According to "FIDO AppID and Facet Specification",
https://somehost.somedomain/appID should be ok. But chrome and firefox
just plain refuse to anything with it. 

If somebody reads this: is there any reference implementation of the js
client part? Right now I'm using the u2f-api.js for testing which I got
from the yubico test site. I don't want to do this part myself because,
as a programming language, js sucks big time.